ci: rename test folder so we dont clash rolename
We are unable to completely control ANSIBLE_ROLES_PATH when running
molecule. All we can do is append to the end it seems. So what was
happening is that community.zabbix/molecule/zabbix_agent was being
discovered first as an empty role (I think). By changing this dir
name we are able to let Molecule/Ansible eventually find the correct
location for the role.